Add files via upload
Here are some screenshots of Tableau visualizations I created for my team's House Price Index Analysis Hackathon. The prompt for this hackathon was: "How and why is the HPI different among different regions in the United States?". We had two weeks to complete this project and presented to the Mortage Analytics team upon completion. Our team focused on historical changes of HPI in Florida and Texas(states in the Truist footprint) and explored unemployment, civilian labor population, and median income data to generate insights. Ultimately, our team won for having the best presentation, and insights.